Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
opentracing==2.4.0
Flask-OpenTracing==1.1.0
jaeger-client==4.3.0
prometheus-flask-exporter==0.18.2
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-sdk
Flask-PyMongo==2.3.0